Major rework, streaming gRPC — WIP
- Remove OwnedHashTreeNode
- Add Skipped variant to StoredChunkRef
- Remove `transfer` endpoint from REST API
- Switch to u64 for sizes
- Implement serialization for StoredChunkRef, node's children are
  serialized as a struct with hash and size only, on deserialization
  they are rebuilt as a Skipped node and need to be filled from storage
  before they can be used
- Make diff_tree() streaming, the new one should work nicely with gRPC
  while copying less data than before. It's far from zero-copy but
  certainly an improvement

+++ core/src/item.rs
@@ -128,14 +128,14 @@ impl Item {
     /// Recompute total size of the item
     /// Computed as the sum of the sizes of the chunks
     #[must_use]
-    pub fn recompute_size(&self) -> u32 {
+    pub fn recompute_size(&self) -> u64 {
         // useful?
         self.chunks.iter().map(|x| x.size).sum()
     }
 
     /// Total size of the item
     #[must_use]
-    pub fn size(&self) -> u32 {
+    pub fn size(&self) -> u64 {
         self.metadata.size()
     }
 
@@ -250,7 +250,7 @@ pub mod tests {
         let data = Bytes::from_iter([value; CHUNK_SIZE]);
         let chunk = ChunkInfo {
             hash: hash(&data),
-            size: CHUNK_SIZE as u32,
+            size: CHUNK_SIZE as u64,
         };
         Item::make(
             "name".to_string(),
